i think were done

+++ third year project/Services/SineVoice.cs
@@ -10,11 +10,11 @@ namespace third_year_project.Services
     {
         public double phase;
         public double phaseIncrement;
-        public float amplitude;
+        public double amplitude;
         public long remainingSamples;
         public Note? note;
 
-        public SineVoice(double _frequency, float _amplitude, int _sampleRate, long _durationSamples)
+        public SineVoice(double _frequency, double _amplitude, int _sampleRate, long _durationSamples)
         {
             phase = 0;
             phaseIncrement = 2.0 * Math.PI * _frequency / _sampleRate;
@@ -22,7 +22,7 @@ namespace third_year_project.Services
             remainingSamples = _durationSamples;
         }
 
-        public SineVoice(double _frequency, float _amplitude, int _sampleRate, long _durationSamples, Note _note)
+        public SineVoice(double _frequency, double _amplitude, int _sampleRate, long _durationSamples, Note _note)
         {
             phase = 0;
             phaseIncrement = 2.0 * Math.PI * _frequency / _sampleRate;
